pyproject.toml
[tool.poetry]
name = "curse-work-4oop"
version = "0.1.0"
description = ""
authors = ["zixeel <strpan549@gmail.com >"]
readme = "README.md"
packages = [{include = "curse_work_4oop"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
